requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
xgboost
streamlit
psycopg2-binary
pickle-mixin
